+++ analog/analyzer.py
@@ -5,6 +5,12 @@ import datetime
 import re
 import time
 
+try:
+    from datetime import timezone
+    UTC = timezone.utc
+except ImportError:
+    from pytz import UTC
+
 from analog.exceptions import MissingArgumentError
 from analog.formats import LogFormat
 from analog.report import Report
@@ -48,8 +54,7 @@ class Analyzer:
         self._pathconf = paths
 
         self._now = datetime.datetime.utcnow()
-        self._now = self._now.replace(second=0, microsecond=0,
-                                      tzinfo=datetime.timezone.utc)
+        self._now = self._now.replace(second=0, microsecond=0, tzinfo=UTC)
         self._min_time = self._now - datetime.timedelta(minutes=max_age)
 
     def _monitor_path(self, path):