+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 0.1.0 - unreleased
 ------------------
 
+* Install pytz on Python <= 3.2 for UTC object. Else use datetime.timezone.
+
 * Add tox environment for py2.7 and py3.3 testing.
 
 * Initial implementation of log analyzer and report object.